pyproject.toml
[tool.poetry]
name = "shraga-common"
version = "0.1.0"
description = ""
authors = ["lizozom <liza.katz@bigdataboutique.com>"]
readme = "README.md"
packages = [{include = "shraga_common"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.2"
boto3 = "^1.37.10"
pydash = "^8.0.3"
opensearch-py = "^2.6.0"
langchain-core = "^0.2.29"
langchain = "^0.2.12"
pandas = "^2.2.2"
openai = "1.57.4"
snowflake-connector-python = "^3.11.0"
elasticsearch = "^8.14.0"
geopandas = "^1.0.1"
googlemaps = "^4.10.0"
vertexai = "^1.60.0"
faiss-cpu = "^1.8.0.post1"
pydantic = "^2.10.6"
typing-extensions = "^4.12.2"
colorlog = "^6.9.0"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
user-agents = "^2.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
autoflake = "^2.3.1"
pre-commit = "^3.7.1"
ipykernel = "^6.29.5"
selenium = "^4.27.1"
chromedriver-binary = "130.0.6723.116"
pylint = "^3.3.4"


